Todo listo menos monitores

+++ Inventario5/Models/Portatil.cs
@@ -8,7 +8,7 @@ namespace Inventario5.Models
     public partial class Portatil
     {
         public int Id { get; set; }
-        public byte? TamanoPantalla { get; set; }
+        public float? TamanoPantalla { get; set; }
 
         public virtual Ordenador IdNavigation { get; set; }
     }
